Merge the Qt frontend release (pre-release) into main for 2.1.0.

Conflicts resolved by taking pre-release's versions, verified equivalent:
- .github/workflows/release-windows.yml: main's copy was a verbatim older
  snapshot ('No behavior change'); pre-release carries the current workflow.
- README.md: pre-release's Qt rewrite already covers main's Aug-31 additions
  (native/not-a-web-player, bit-perfect DAC, no EQ/DSP, TDL/SPTFY/DZR, offline
  locks) in its dedicated 'What QBZ is and what it is not' section.

cc

# Conflicts:
#	.github/workflows/release-windows.yml
#	README.md

+++ crates/qbz-reco/src/store.rs
@@ -221,7 +221,7 @@ impl ArtistVectorStore {
             .execute(
                 "INSERT OR REPLACE INTO vector_metadata (artist_idx, updated_at, nnz)
                  VALUES (?1, ?2, ?3)",
-                params![artist_idx, now, vector.nnz()],
+                params![artist_idx, now, vector.nnz() as i64],
             )
             .map_err(|e| format!("Failed to update metadata: {}", e))?;
